page_imagesの最新をcache

+++ batch.sh
@@ -90,8 +90,18 @@ fi
 cp -r ./work/cache/* ./projects/${pj}/cache || exit 1
 if [ $stop = 'voice' ]; then exit 0; fi
 echo '# pdf2png'
-mkdir ./work/page_images
-docker run --rm -v $PWD/work:/work gkmr/pdf-tools /bin/sh -c "pdftocairo -png -r 200 /work/novel.pdf /work/page_images/novel" || exit 1
+hash=`docker run --rm -v $PWD/work:/work lp-python /bin/sh -c "tail -n +2 /work/novel.log | md5sum | cut -d ' ' -f 1"` || exit 1
+if [ -e ./projects/${pj}/cache/page_images/${hash}.zip ]; then
+  echo '## exists cache'
+  unzip ./projects/${pj}/cache/page_images/${hash}.zip
+else
+  echo '## processing...'
+  mkdir -p ./work/page_images || exit 1
+  docker run --rm -v $PWD/work:/work gkmr/pdf-tools /bin/sh -c "pdftocairo -png -r 200 /work/novel.pdf /work/page_images/novel" || exit 1
+  rm -rf ./projects/${pj}/cache/page_images/ || exit 1
+  mkdir ./projects/${pj}/cache/page_images/ || exit 1
+  zip -r "./projects/${pj}/cache/page_images/${hash}.zip" ./work/page_images || exit 1
+fi
 echo '# build_timekeeper'
 docker run --rm -v $PWD/work:/work lp-python /bin/sh -c "python -u build_timekeeper.py" || exit 1
 if [ $stop = 'timekeeper' ]; then exit 0; fi
